nvim: term: fix issues with excessive escaping
The `:split` trick fixed one issue (`%` not expanding) and replaced it
with another: the need to excessively escape slashes due to multiple
layers of quoting. This led to things like `:Rg 'foo\\\\('` being
necessary to perform `rg 'foo\('`.

This solution solves the `%` expansion issue and allows you to get away
with only escaping as much as is absolutely necessary
(i.e. `:Rg 'foo\('` -> `rg 'foo\('`)